refactor(python): move out TOIF related structs to trezorlib.toif
